+++ src/index.ts
@@ -1,7 +1,7 @@
 import type { MockedRequest, SetupWorker } from 'msw';
 import type { SetupServer } from 'msw/node';
 import { pathToRegexp } from 'path-to-regexp';
-import { defaultRequestMapper } from './defaultRequestMapper';
+import { defaultRequestLogger } from './defaultRequestLogger';
 import { makeErrorMessage } from './makeErrorMessage';
 
 type RequestLog = { req: MockedRequest; record: Record<string, unknown> };
@@ -12,21 +12,20 @@ type RequestLog = { req: MockedRequest; record: Record<string, unknown> };
 function createMSWInspector<FunctionMock extends Function>({
   mockSetup,
   mockFactory,
-  requestMapper = defaultRequestMapper,
+  requestLogger = defaultRequestLogger,
 }: {
   mockSetup: SetupServer | SetupWorker;
   mockFactory: () => FunctionMock;
-  requestMapper?: (req: MockedRequest) => Promise<Record<string, unknown>>;
+  requestLogger?: (req: MockedRequest) => Promise<Record<string, unknown>>;
 }) {
   // Store network requests by url
-  const requestLog = new Map<string, RequestLog[]>();
+  const requestLogs = new Map<string, RequestLog[]>();
 
   async function logRequest(req: MockedRequest): Promise<void> {
     const { href } = req.url;
-    const reqs = requestLog.get(href) || [];
-    // @NOTE we need to create the request record now since we can deserialize body only once
-    reqs.push({ req, record: await requestMapper(req) });
-    requestLog.set(href, reqs);
+    const currentRequestLogs = requestLogs.get(href) || [];
+    currentRequestLogs.push({ req, record: await requestLogger(req) });
+    requestLogs.set(href, currentRequestLogs);
   }
 
   return {
@@ -39,7 +38,7 @@ function createMSWInspector<FunctionMock extends Function>({
      */
     async getRequests(path: string): Promise<FunctionMock> {
       const matches: RequestLog[] = [];
-      requestLog.forEach((requests, requestHref) => {
+      requestLogs.forEach((requests, requestHref) => {
         const requestsURL = new URL(requestHref);
         let pathURL: URL;
         try {
@@ -48,7 +47,7 @@ function createMSWInspector<FunctionMock extends Function>({
           throw new Error(
             makeErrorMessage({
               message: `Provided path is invalid: ${path}`,
-              requestLog,
+              requestLogs,
             }),
           );
         }
@@ -67,7 +66,7 @@ function createMSWInspector<FunctionMock extends Function>({
         throw new Error(
           makeErrorMessage({
             message: `Cannot find a matching requests for path: ${path}`,
-            requestLog,
+            requestLogs,
           }),
         );
       }
@@ -93,7 +92,7 @@ function createMSWInspector<FunctionMock extends Function>({
      * Clear msw spy call log. Call it before every single test execution
      */
     clear() {
-      requestLog.clear();
+      requestLogs.clear();
       return this;
     },
 
@@ -109,4 +108,4 @@ function createMSWInspector<FunctionMock extends Function>({
 }
 
 export type MswInspector = ReturnType<typeof createMSWInspector>;
-export { createMSWInspector, defaultRequestMapper };
+export { createMSWInspector, defaultRequestLogger };