+++ src/makeErrorMessage.ts
@@ -1,11 +1,11 @@
 export function makeErrorMessage({
   message,
-  requestLog,
+  requestLogs,
 }: {
   message: string;
-  requestLog: Map<string, unknown>;
+  requestLogs: Map<string, unknown>;
 }): string {
-  const availablePaths = Array.from(requestLog.keys());
+  const availablePaths = Array.from(requestLogs.keys());
   return `[msw-inspector] ${message}. Intercepted requests paths are:\n\n${availablePaths.join(
     '\n',
   )}`;